connections movable.
Connections move with the devices the connect to. There are a couple of bugs: one causes other connections than the one currently moving to disappear until the deivec the disappearing connection is connected to is moved again. Also, app throws an exception if you attempt to add new connection from a newly added device after the first connection ina topology is made; you must select a device that already has a connection and make the connection from that already 'initialized' device. Problem is in finishConnection function on line 307. Will fix later, going to bed :smile: