Connection bug fixed.
Connections drawn properly when moving. problem was not in finishConnection, it was the wierd 'for' construct in 'dragDevice', it was changing the position of other connections beside the one that should be moving because there was no condition to select only the moving connection. I restructured the 'for' stack (which i hope to redo one day soon to come up with something better, if such exists) to ensure proper comparison of the connection indexes to prevent the wrong ones having their positions modified.