Connections now removed with devices
When a device is deleted, its connection is deleted as well. This implementation is flawed as it leaves the connection in place in the connections array to maintain the indexes of the remaining ones. A function will be constructed later that actually removes the connection and updates the values in each remaining devices connectionIndex array. Right now it removes connections by setting their state to invalid so they aren't drawn and also removes the reference to the deleted device from the remaining one's 'connectedTo' array.